refactor: 라우터 구조 개선 및 라우트 그룹별 파일 분리
